Hyperskill
Residents of the country of Malevia often experiment with the plan of their rooms. Rooms can be triangular, rectangular, and round. To quickly calculate the floorage it is required to write a program that gets the type of the room shape and the relevant parameters as input. The program should output the area of the resulting room.

The value of 3.14 is used instead of the number π in Malevia.

Hint: Math library may help you to solve this task.