Create alphabeta.java
Our friends Alpha and Beta found a magical treasure of Asgard. It consists of  piles of gold coins. It is magical since if anyone tries to take a pile of  coins, all the other piles of exactly  coins (if they exist) disappear.

Alpha and Beta only have one turn each to choose a pile for themselves starting with Alpha. In one turn, a complete pile of gold coins can be chosen and since our friends are smart they will choose the pile with the maximum coins.

Find the number of coins Beta will get in his turn.